+++ src/otx/cli/cli.py
@@ -220,6 +220,7 @@ class OTXCLI:
             "export": device_kwargs,
             "optimize": {"datamodule"}.union(device_kwargs),
             "explain": {"datamodule"}.union(device_kwargs),
+            "benchmark": device_kwargs,
         }
 
     def add_subcommands(self) -> None: