+++ src/otx/core/model/classification.py
@@ -63,6 +63,7 @@ class OTXMulticlassClsModel(OTXModel[MulticlassClsBatchDataEntity, MulticlassCls
             metric=metric,
             torch_compile=torch_compile,
         )
+        self.image_size = (1, 3, 224, 224)
 
     @property
     def _export_parameters(self) -> TaskLevelExportParameters:
@@ -86,6 +87,12 @@ class OTXMulticlassClsModel(OTXModel[MulticlassClsBatchDataEntity, MulticlassCls
             "target": target,
         }
 
+    def get_dummy_input(self, batch_size: int = 1) -> MulticlassClsBatchDataEntity:
+        """Returns a dummy input for classification model."""
+        images = [torch.rand(*self.image_size[1:]) for _ in range(batch_size)]
+        labels = [torch.LongTensor([0])] * batch_size
+        return MulticlassClsBatchDataEntity(batch_size, images, [], labels=labels)
+
 
 class MMPretrainMulticlassClsModel(OTXMulticlassClsModel):
     """Multi-class Classification model compatible for MMPretrain.
@@ -107,7 +114,6 @@ class MMPretrainMulticlassClsModel(OTXMulticlassClsModel):
         config = inplace_num_classes(cfg=config, num_classes=self._dispatch_label_info(label_info).num_classes)
         self.config = config
         self.load_from = config.pop("load_from", None)
-        self.image_size = (1, 3, 224, 224)
         super().__init__(
             label_info=label_info,
             optimizer=optimizer,
@@ -255,6 +261,7 @@ class OTXMultilabelClsModel(OTXModel[MultilabelClsBatchDataEntity, MultilabelCls
             metric=metric,
             torch_compile=torch_compile,
         )
+        self.image_size = (1, 3, 224, 224)
 
     @property
     def _export_parameters(self) -> TaskLevelExportParameters:
@@ -281,6 +288,12 @@ class OTXMultilabelClsModel(OTXModel[MultilabelClsBatchDataEntity, MultilabelCls
         """Model forward function used for the model tracing during model exportation."""
         return self.model.forward(image, mode="tensor")
 
+    def get_dummy_input(self, batch_size: int = 1) -> MultilabelClsBatchDataEntity:
+        """Returns a dummy input for classification OV model."""
+        images = [torch.rand(*self.image_size[1:]) for _ in range(batch_size)]
+        labels = [torch.LongTensor([0])] * batch_size
+        return MultilabelClsBatchDataEntity(batch_size, images, [], labels=labels)
+
 
 class MMPretrainMultilabelClsModel(OTXMultilabelClsModel):
     """Multi-label Classification model compatible for MMPretrain.
@@ -302,7 +315,6 @@ class MMPretrainMultilabelClsModel(OTXMultilabelClsModel):
         config = inplace_num_classes(cfg=config, num_classes=self._dispatch_label_info(label_info).num_classes)
         self.config = config
         self.load_from = config.pop("load_from", None)
-        self.image_size = (1, 3, 224, 224)
         super().__init__(
             label_info=label_info,
             optimizer=optimizer,
@@ -444,6 +456,7 @@ class OTXHlabelClsModel(OTXModel[HlabelClsBatchDataEntity, HlabelClsBatchPredEnt
             metric=metric,
             torch_compile=torch_compile,
         )
+        self.image_size = (1, 3, 224, 224)
 
     @property
     def _export_parameters(self) -> TaskLevelExportParameters:
@@ -481,6 +494,12 @@ class OTXHlabelClsModel(OTXModel[HlabelClsBatchDataEntity, HlabelClsBatchPredEnt
 
         return label_info
 
+    def get_dummy_input(self, batch_size: int = 1) -> HlabelClsBatchDataEntity:
+        """Returns a dummy input for classification OV model."""
+        images = [torch.rand(*self.image_size[1:]) for _ in range(batch_size)]
+        labels = [torch.LongTensor([0])] * batch_size
+        return HlabelClsBatchDataEntity(batch_size, images, [], labels=labels)
+
 
 class MMPretrainHlabelClsModel(OTXHlabelClsModel):
     """H-label Classification model compatible for MMPretrain.
@@ -509,7 +528,6 @@ class MMPretrainHlabelClsModel(OTXHlabelClsModel):
 
         self.config = config
         self.load_from = config.pop("load_from", None)
-        self.image_size = (1, 3, 224, 224)
         super().__init__(
             label_info=label_info,
             optimizer=optimizer,